backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.110.1
uvicorn==0.25.0
starlette==0.37.2
pydantic==2.12.3
pydantic_core==2.41.4

# Authentication & Security
PyJWT==2.10.1
python-jose==3.5.0
bcrypt==4.1.3
passlib==1.7.4
cryptography==43.0.0
ecdsa==0.19.1
rsa==4.9.1
pyasn1==0.6.1

# Database
motor==3.3.1
pymongo==4.5.0
dnspython==2.7.0

# Environment & Configuration
python-dotenv==1.1.1
python-multipart==0.0.20

# HTTP & Requests
httpx==0.27.0
requests==2.32.5
urllib3==2.5.0
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
httpcore==1.0.2

# Async & Utilities
anyio==4.11.0
sniffio==1.3.1
h11==0.14.0
click==8.1.8
typing_extensions==4.15.0

# Email validation
email-validator==2.3.0

# Date/Time
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2

# Other utilities
six==1.17.0
cffi==2.0.0
pycparser==2.23
annotated-types==0.7.0
